# backend/cmd/server/main.go
package main

import (
	"context"
	"fmt"
	"log"
	"log/slog"
	"net"
	"net/http"
	"os"
	"os/signal"
	"syscall"
	"time"

	"github.com/OpenNSW/agency/backend/internal/application"
	"github.com/OpenNSW/agency/backend/internal/authn"
	"github.com/OpenNSW/agency/backend/internal/certificate"
	"github.com/OpenNSW/agency/backend/internal/consignment"
	"github.com/OpenNSW/agency/backend/internal/feedback"
	"github.com/OpenNSW/agency/backend/internal/logging"
	"github.com/OpenNSW/agency/backend/internal/nswclient"
	"github.com/OpenNSW/agency/backend/internal/rbac"
	"github.com/OpenNSW/agency/backend/internal/scopes"
	"github.com/OpenNSW/agency/backend/internal/storage"
	"github.com/OpenNSW/agency/backend/internal/user"
	"github.com/OpenNSW/agency/backend/internal/web"
	"github.com/OpenNSW/core/artifact"
	"github.com/OpenNSW/core/artifact/loaders"
	"github.com/OpenNSW/core/authz"
	"github.com/OpenNSW/core/trace"
)

func main() {
	logging.ConfigureLogging(os.Stdout)

	cfg, err := LoadConfig()
	if err != nil {
		log.Fatalf("FATAL: failed to load configuration: %v", err)
	}

	dbTarget := cfg.DB.SQLite.Path
	if cfg.DB.Driver == "postgres" {
		dbTarget = net.JoinHostPort(cfg.DB.Postgres.Host, cfg.DB.Postgres.Port) + "/" + cfg.DB.Postgres.Name
	}

	slog.Info("NSW Agency service configuration",
		"db_driver", cfg.DB.Driver,
		"db_target", dbTarget,
		"port", cfg.Port,
	)

	// Initialize database store
	store, err := application.NewApplicationStore(cfg.DB)
	if err != nil {
		log.Fatalf("failed to create application store: %v", err)
	}

	// Initialize user store
	userStore, err := user.NewUserStore(cfg.DB)
	if err != nil {
		log.Fatalf("failed to create user store: %v", err)
	}
	defer func() {
		if err := userStore.Close(); err != nil {
			slog.Error("failed to close user store", "error", err)
		}
	}()

	// Initialize auth manager, resolving callers to seeded user records.
	authManager, err := authn.NewManager(&userProfileAdapter{store: userStore}, cfg.Authn)
	if err != nil {
		log.Fatalf("failed to initialize auth manager: %v", err)
	}
	defer func() {
		if err := authManager.Close(); err != nil {
			slog.Error("failed to close auth manager", "error", err)
		}
	}()

	// Authorizer: gates routes by the OAuth2 scopes carried on the token. The
	// extractor bridges internal/authn's Principal into authz.Principal via
	// authzPrincipal — authz imports nothing from internal/authn.
	authzr, err := authz.New(func(ctx context.Context) (authz.Principal, bool) {
		p, ok := authn.FromContext(ctx)
		if !ok || p == nil {
			return nil, false
		}
		return authzPrincipal{p}, true
	})
	if err != nil {
		log.Fatalf("failed to initialize authorizer: %v", err)
	}

	// NSW client: anti-corruption layer that owns the NSW HTTP transport,
	// OAuth2 credentials, and wire protocol.
	nswClient := nswclient.New(cfg.NSW)

	// Bound startup on the remote artifact store: a slow or unreachable backend
	// (GitHub / S3) must fail fast with a clear error rather than hang the boot.
	initCtx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
	defer cancel()

	artifactLoader, err := loaders.New(initCtx, cfg.ArtifactLoader)
	if err != nil {
		log.Fatalf("failed to initialize artifact loader: %v", err)
	}

	artifactRegistry := artifact.NewRegistry(artifactLoader)

	manifestCfg, err := artifact.LoadManifest(initCtx, artifactLoader)
	if err != nil {
		log.Fatalf("failed to load artifact manifest: %v", err)
	}
	if err := artifact.RegisterFromConfig(artifactRegistry, manifestCfg); err != nil {
		log.Fatalf("failed to register artifacts from manifest: %v", err)
	}

	// Initialize RBAC Service and Middleware
	roleService := rbac.NewRoleService(store.DB())
	rbacMiddleware := rbac.NewMiddleware(roleService, store, artifactRegistry)

	// Initialize consignment service (read-only summaries over the shared DB connection)
	consignmentStore := consignment.NewConsignmentStore(store.DB())
	consignmentHandler := consignment.NewHandler(consignment.NewService(consignmentStore, nswClient))

	// Initialize Agency service
	service := application.NewService(store, artifactRegistry, nswClient, roleService)
	defer func() {
		if err := service.Close(); err != nil {
			slog.Error("failed to close service", "error", err)
		}
	}()

	// Initialize handlers
	handler, err := application.NewHandler(service, cfg.MaxRequestBytes)
	if err != nil {
		slog.Error("failed to create Agency handler", "error", err)
		return
	}

	profileSvc := user.NewProfileService(roleService)
	profileHandler := user.NewProfileHandler(profileSvc)

	// Initialize storage handler (delegates NSW backend calls to nswClient)
	storageHandler, err := storage.NewHandler(nswClient, cfg.MaxRequestBytes)
	if err != nil {
		slog.Error("failed to create storage handler", "error", err)
		return
	}

	feedbackHandler, err := feedback.NewHandler(service, cfg.MaxRequestBytes)
	if err != nil {
		slog.Error("failed to create feedback handler", "error", err)
		return
	}

	// Initialize certificate handler (populates gohtml templates fetched from the artifact registry)
	certificateService := certificate.NewService(artifactRegistry, service)
	certificateHandler, err := certificate.NewHandler(certificateService, service, cfg.MaxRequestBytes)
	if err != nil {
		slog.Error("failed to create certificate handler", "error", err)
		return
	}

	// Set up HTTP routes
	mux := http.NewServeMux()
	// Health check
	mux.HandleFunc("GET /health", handler.HandleHealth)

	// Shared auth middleware: requires a valid IdP token whose client_id is in
	// AUTH_CLIENT_IDS and aud=AGENCY_API.
	protect := authManager.RequireAuthMiddleware()

	// withScope returns middleware requiring the given OAuth2 scope; compose
	// after protect so the authn principal is already injected when the scope
	// check runs.
	withScope := authzr.RequireScope

	// Endpoint for services to inject data (service-to-service M2M). Protected by
	// the same auth middleware; the NSW->Agency M2M client (e.g. NSW_TO_NPQS) is
	// whitelisted in AUTH_CLIENT_IDS so NSW core authenticates with its
	// client_credentials token.
	mux.Handle("POST /api/v1/inject", protect(withScope(scopes.ApplicationInject)(http.HandlerFunc(handler.HandleInjectData))))

	// Endpoints for UI to fetch and manage applications (protected by JIT user auth)
	mux.Handle("GET /api/v1/consignments", protect(withScope(scopes.ConsignmentRead)(http.HandlerFunc(consignmentHandler.HandleGetConsignments))))
	mux.Handle("GET /api/v1/consignments/{consignmentId}", protect(withScope(scopes.ConsignmentRead)(http.HandlerFunc(consignmentHandler.HandleGetConsignment))))
	mux.Handle("GET /api/v1/applications", protect(withScope(scopes.ApplicationRead)(http.HandlerFunc(handler.HandleGetApplications))))
	mux.Handle("GET /api/v1/users/me", protect(withScope(scopes.ProfileRead)(http.HandlerFunc(profileHandler.HandleMe))))
	mux.Handle("GET /api/v1/applications/{taskId}", protect(withScope(scopes.ApplicationRead)(rbacMiddleware.RequireAction("VIEW")(http.HandlerFunc(handler.HandleGetApplication)))))
	mux.Handle("POST /api/v1/applications/{taskId}/review", protect(withScope(scopes.ApplicationReview)(rbacMiddleware.RequireAction("REVIEW")(http.HandlerFunc(handler.HandleReviewApplication)))))
	mux.Handle("POST /api/v1/applications/{taskId}/feedback", protect(withScope(scopes.ApplicationFeedback)(rbacMiddleware.RequireAction("FEEDBACK")(http.HandlerFunc(feedbackHandler.HandleFeedback)))))
	mux.Handle("POST /api/v1/applications/{taskId}/claim", protect(withScope(scopes.ApplicationReview)(rbacMiddleware.RequireAction("REVIEW")(http.HandlerFunc(handler.HandleClaimApplication)))))
	mux.Handle("POST /api/v1/applications/{taskId}/release", protect(withScope(scopes.ApplicationReview)(rbacMiddleware.RequireAction("REVIEW")(http.HandlerFunc(handler.HandleReleaseApplication)))))
	mux.Handle("POST /api/v1/storage", protect(withScope(scopes.StorageWrite)(http.HandlerFunc(storageHandler.HandleCreateUpload))))
	mux.Handle("GET /api/v1/storage/{key}", protect(withScope(scopes.StorageRead)(http.HandlerFunc(storageHandler.HandleGetUploadURL))))
	mux.Handle("POST /api/v1/applications/{taskId}/certificate", protect(withScope(scopes.ApplicationReview)(rbacMiddleware.RequireAction("REVIEW")(http.HandlerFunc(certificateHandler.HandleGenerate)))))

	// Serve the built officer-portal SPA from this same process. The "/" pattern
	// is the most general match, so the specific API, /health and /runtime-env.js
	// routes take precedence. /runtime-env.js exposes cfg.Web.Runtime
	// (window.__APP_CONFIG__) so the SPA reads config synchronously. cfg.Web.Dir
	// is relative to the working dir (/app/web in the image); when it is absent
	// (e.g. local API-only dev where the frontend runs via its own dev server),
	// skip serving rather than failing.
	if spa, err := web.NewHandler(cfg.Web); err != nil {
		slog.Warn("frontend not served", "asset_dir", cfg.Web.Dir, "error", err)
	} else {
		if err := cfg.Web.Validate(); err != nil {
			log.Fatalf("FATAL: serving the frontend but its config is invalid: %v", err)
		}
		mux.Handle("GET /runtime-env.js", http.HandlerFunc(spa.ServeRuntimeEnv))
		mux.Handle("GET /", http.HandlerFunc(spa.ServeSPA))
		slog.Info("serving officer portal alongside the API", "asset_dir", cfg.Web.Dir)
	}

	// Set up graceful shutdown
	serverAddr := fmt.Sprintf(":%s", cfg.Port)

	// CORS middleware
	allowAll := len(cfg.AllowedOrigins) == 1 && cfg.AllowedOrigins[0] == "*"
	allowedSet := make(map[string]struct{}, len(cfg.AllowedOrigins))
	for _, o := range cfg.AllowedOrigins {
		allowedSet[o] = struct{}{}
	}

	corsHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		origin := r.Header.Get("Origin")
		if allowAll {
			w.Header().Set("Access-Control-Allow-Origin", "*")
		} else if _, ok := allowedSet[origin]; ok {
			w.Header().Set("Access-Control-Allow-Origin", origin)
			w.Header().Set("Vary", "Origin")
		}
		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
		if r.Method == "OPTIONS" {
			w.WriteHeader(http.StatusOK)
			return
		}
		mux.ServeHTTP(w, r)
	})

	server := &http.Server{
		Addr:              serverAddr,
		Handler:           trace.TraceMiddleware(corsHandler),
		ReadHeaderTimeout: cfg.ReadHeaderTimeout,
		ReadTimeout:       cfg.ReadTimeout,
		WriteTimeout:      cfg.WriteTimeout,
		IdleTimeout:       cfg.IdleTimeout,
	}

	// Channel to listen for interrupt signals
	quit := make(chan os.Signal, 1)
	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)

	// Start server in a goroutine
	go func() {
		slog.Info("starting Agency service", "port", cfg.Port)
		if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			slog.Error("failed to start server", "error", err)
			quit <- syscall.SIGTERM
		}
	}()

	// Wait for interrupt signal
	<-quit
	slog.Info("shutting down Agency service...")

	// Create a context with timeout for graceful shutdown
	shutdownCtx, shutdownCancel := context.WithTimeout(context.Background(), 30*time.Second)
	defer shutdownCancel()

	// Attempt graceful shutdown of HTTP server
	if err := server.Shutdown(shutdownCtx); err != nil {
		slog.Error("server forced to shutdown", "error", err)
	} else {
		slog.Info("server gracefully stopped")
	}

	slog.Info("NSW Agency service stopped")
}